Remove strymon_runtime crate
This finishes the removal of the strymon_runtime monolith and splits
its remaining functionality into a bunch of smaller crates:

  strymon_cli: The `strymon` command-line interface

  strymon_job: Client library, to be used to write Strymon jobs
  strymon_coordinator: The implementation of the coordinator and catalog
  strymon_executor: Executor and job spawn logic

  strymon_rpc: The wire-protocols used between the different components
  strymon_model: Common data-types used by different parts of the system
  strymon_communication: Low-level communication primitives
