+++ src/strymon_cli/src/errors.rs
