+++ src/strymon_cli/src/submit/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2017 ETH Zurich. All rights reserved.
+// Copyright 2018 ETH Zurich. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
 // http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
@@ -6,117 +6,24 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::io;
+
 use std::env;
 use std::path::Path;
 use std::ffi::OsStr;
-use std::process::{Command, Stdio};
 
-use serde_json::{Value, Deserializer};
 use clap::{App, AppSettings, Arg, ArgGroup, ArgMatches, SubCommand};
-use log::Level;
+
 
 use strymon_communication::Network;
-use strymon_runtime::submit::Submitter;
 use strymon_model::{QueryProgram, QueryId, ExecutionFormat, Executor, ExecutorId};
 use strymon_rpc::coordinator::Placement;
 
 use errors::*;
 
-fn build_binary(path: &Path, args: &ArgMatches) -> Result<String> {
-    let mut cargo = Command::new("cargo");
-    cargo.arg("build")
-        .args(&["--message-format", "json"]);
+pub use self::submitter::Submitter;
 
-    // translate project directory to manifest path
-    let manifest = path.join("Cargo.toml");
-    cargo.arg("--manifest-path").arg(manifest);
-
-    // pass down custom cargo flags
-    if !args.is_present("--debug") {
-        cargo.arg("--release");
-    }
-
-    if args.is_present("no-default-features") {
-        cargo.arg("--no-default-features");
-    }
-    if args.is_present("all-features") {
-        cargo.arg("--all-features");
-    } else if let Some(list) = args.values_of("features") {
-        cargo.arg("--features").arg(list.collect::<Vec<_>>().join(" "));
-    };
-
-    if let Some(name) = args.value_of("bin") {
-        cargo.args(&["--bin", name]);
-    } else if let Some(name) = args.value_of("example") {
-        cargo.args(&["--example", name]);
-    }
-
-    // captures stderr and stdout
-    cargo.stdin(Stdio::null());
-    cargo.stderr(Stdio::piped());
-    cargo.stdout(Stdio::piped());
-
-    // list of all compiled binaries, we need exactly one
-    let mut binaries: Vec<String> = vec![];
-
-    info!("running: `{:?}`", cargo);
-    eprintln!("Building job binary with `cargo build` (this will take a while)");
-
-    // spawn cargo as a child process
-    let mut child = cargo.spawn()?;
-
-    // TODO(swicki): Have proper deserialize struct for parsing the messages
-    let stream = Deserializer::from_reader(child.stdout.take().unwrap()).into_iter::<Value>();
-    for result in stream {
-        let msg = result?;
-        match msg["reason"].as_str().ok_or("missing reason field in cargo message")? {
-            "compiler-message" => {
-                let rustc_msg = msg.get("message").ok_or("missing compiler message")?;
-                let level = match rustc_msg["level"].as_str().ok_or("unable to parse message level")? {
-                    "note" | "help" => Level::Info,
-                    "warning" => Level::Warn,
-                    "error" => Level::Error,
-                    _ => Level::Debug,
-                };
-                log!(level, "cargo: {}", msg);
-            },
-            "compiler-artifact" => {
-                if let Value::String(ref crate_name) = msg["package_id"] {
-                    info!("compiled crate: {}", crate_name);
-                }
-                // TODO(swicki): When do artifacts have more than one type/kind?
-                let ref crate_type = msg["target"]["crate_types"][0];
-                let ref kind = msg["target"]["kind"][0];
-                if crate_type == "bin" && (kind == "bin" || kind == "example") {
-                    let file = msg["filenames"][0].as_str().ok_or("missing filename")?;
-                    binaries.push(file.to_owned());
-                }
-            }
-            _ => {
-                debug!("unhandled cargo message: {}", msg);
-            }
-        };
-    }
-
-    // cargo has closed stdout, wait for it to finish
-    let status = child.wait()?;
-    if !status.success() {
-        io::copy(child.stderr.as_mut().unwrap(), &mut io::stderr())?;
-        bail!("Cargo failed with exit code {}", status.code().unwrap_or(-1));
-    }
-
-    if binaries.is_empty() {
-        bail!("Cargo did not produce any binaries, a `bin` target must be available");
-    }
-
-    if binaries.len() > 1 {
-        bail!("Multiple binaries for this project, please specify which one to submit using `--bin` or `--example`");
-    }
-
-    let binary = binaries.pop().unwrap();
-    Ok(binary)
-}
+mod submitter;
+mod build;
 
 fn parse_placement(args: &ArgMatches, executors: Vec<Executor>) -> Result<Placement> {
     fn parse_err(arg: &str) -> String {
@@ -379,7 +286,7 @@ pub fn usage<'a, 'b>() -> App<'a, 'b> {
 
 pub fn main(args: &ArgMatches) -> Result<()> {
     let binary = if let Some(path) = args.value_of("path").map(Path::new) {
-        build_binary(path, args)?
+        build::binary(path, args)?
     } else {
         args.value_of("binary-path").expect("no binary specified").to_owned()
     };
