+++ src/strymon_cli/src/terminate.rs
@@ -10,7 +10,7 @@ use clap::{App, Arg, ArgMatches, SubCommand};
 
 use strymon_model::QueryId;
 use strymon_communication::Network;
-use strymon_runtime::submit::Submitter;
+use super::submit::Submitter;
 
 use errors::*;
