+++ src/strymon_coordinator/Cargo.toml
@@ -1,17 +1,19 @@
 [package]
+name = "strymon_coordinator"
+version = "0.1.0"
 authors = ["Sebastian Wicki <swicki@inf.ethz.ch>"]
-name = "strymon_rpc"
-version = "0.1.2"
 
 [dependencies]
-serde = "1.0"
-serde_derive = "1.0"
-enum-primitive-derive = "0.1"
-num-traits = "0.1"
-typename = "0.1"
+log = "0.4"
+futures = "0.1"
+rand = "0.3"
+tokio-core = "0.1"
 
 [dependencies.strymon_communication]
 path = "../strymon_communication/"
 
 [dependencies.strymon_model]
 path = "../strymon_model/"
+
+[dependencies.strymon_rpc]
+path = "../strymon_rpc/"
