+++ src/strymon_rpc/src/coordinator/mod.rs
@@ -6,6 +6,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+pub mod catalog;
+
 use num_traits::FromPrimitive;
 
 use strymon_model::*;
